+++ src/main/java/com/main_39/Spring/order/controller/OrderController.java
@@ -1,6 +1,6 @@
 package com.main_39.Spring.order.controller;
 
-import com.main_39.Spring.order.OrderMapper;
+import com.main_39.Spring.order.mapper.OrderMapper;
 import com.main_39.Spring.order.controller.dto.OrderPostRequestDto;
 import com.main_39.Spring.order.service.OrderService;
 import com.main_39.Spring.order.entity.Order;
@@ -24,9 +24,7 @@ public class OrderController {
 
     @PostMapping
     public ResponseEntity createOrder(@RequestBody OrderPostRequestDto requestDto) {
-        /**
-         * TODO : OrderPostRequestDto -> Order 변환 필요
-         */
+
         orderService.createOrder(mapper.orderPostRequestDtoToOrder(requestDto));
         return new ResponseEntity ("", HttpStatus.CREATED);
     }
