+++ src/main/java/com/main_39/Spring/order/mapper/OrderMapper.java
@@ -1,4 +1,4 @@
-package com.main_39.Spring.order;
+package com.main_39.Spring.order.mapper;
 
 import com.main_39.Spring.member.entity.Kakao;
 import com.main_39.Spring.menu.Menu;
